+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # ps development version
 
+* New function `ps_cpu_times()` to calculate system CPU times.
+
 * New function `ps_loadavg()` to show the Unix style load average.
 
 # ps 1.5.0